requirements.txt
fastapi
uvicorn
python-multipart
pillow
nudenet
python-jose[cryptography]
